+++ module-extend/module-social/src/main/java/net/linlan/social/third/service/ThirdMemberTokenService.java
@@ -17,7 +17,7 @@
  */
 package net.linlan.social.third.service;
 
-import java.sql.Timestamp;
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
@@ -29,17 +29,26 @@ import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 
 import net.linlan.commons.core.ObjectUtils;
+import net.linlan.commons.core.StringUtils;
 import net.linlan.frame.FrameUserDetails;
 import net.linlan.frame.comm.vo.AppLoginInfo;
 import net.linlan.social.third.dao.ThirdMemberTokenDao;
 import net.linlan.social.third.dto.ThirdMemberTokenDto;
+import net.linlan.social.third.entity.ThirdMember;
 import net.linlan.social.third.entity.ThirdMemberToken;
 import net.linlan.social.third.param.ThirdMemberTokenParam;
+import net.linlan.sys.base.dto.TokenResponse;
+import net.linlan.sys.base.entity.BaseConfigParts;
+import net.linlan.sys.base.entity.BaseUser;
+import net.linlan.sys.base.service.BaseConfigPartsService;
+import net.linlan.sys.base.service.BaseUserService;
+import net.linlan.sys.web.KernelConstant;
+import net.linlan.utils.TokenGenerator;
 
 /**
  *
  * ThirdMemberToken数据域:会员用户令牌服务类
- * 
+ *
  * @author Linlan
  * CreateTime 2025-10-13 16:35:28
  * @version 1.0
@@ -49,8 +58,16 @@ import net.linlan.social.third.param.ThirdMemberTokenParam;
 @Service
 public class ThirdMemberTokenService {
 
+    public final static String     MEMBER_TOKEN_REFRESH_TIME = "MEMBER_TOKEN_REFRESH_TIME";
+
     @Resource
-    private ThirdMemberTokenDao dao;
+    private ThirdMemberTokenDao    dao;
+    @Resource
+    private BaseConfigPartsService baseConfigPartsService;
+    @Resource
+    private ThirdMemberService     thirdMemberService;
+    @Resource
+    private BaseUserService        baseUserService;
 
     /** get the list of entity ThirdMemberToken
      * 列表方法，返回{@link ThirdMemberToken} 列表
@@ -157,14 +174,174 @@ public class ThirdMemberTokenService {
             entity.setForeignId(loginUser.getForeignId());
             entity.setUsername(loginUser.getUsername());
             entity.setLoginIp(loginUser.getLoginIp());
-            entity.setUpdateTime(new Timestamp(System.currentTimeMillis()));
+            entity.setUpdateTime(appLoginInfo.getUpdateTime());
             entity.setExpireTime(appLoginInfo.getExpireTime());
             save(entity);
         } else {
             entity.setToken(appLoginInfo.getToken());
-            entity.setUpdateTime(new Timestamp(System.currentTimeMillis()));
+            entity.setUpdateTime(appLoginInfo.getUpdateTime());
             entity.setExpireTime(appLoginInfo.getExpireTime());
             update(entity);
         }
+    }
+
+    /**
+     * get the ThirdMemberToken by input token
+     *
+     * @param token the input token
+     * @return {@link ThirdMemberToken}
+     */
+    public ThirdMemberToken getByToken(String token) {
+        return dao.getByToken(token);
+    }
+
+    /**
+     * create token of member user by id. 生成用户TOKEN
+     *
+     * @param id       member user id
+     * @param foreignId   site id 或 app id
+     * @param token  token
+     * @param ip       ip address
+     * @return {@link TokenResponse}
+     */
+    public TokenResponse createToken(Long id, String foreignId, String token, String ip) {
+        //当前时间
+        Date now = new Date();
+        //过期时间，默认设置为三十天，后台可调整时间，存储到BASE_CONFIG_PARRTS内，可通过数据库调整
+        //        Date expireTime = DateUtils.addDays(now, KernelConstant.THIRTY_DAY);
+        Long duration = KernelConstant.THIRTY_DAY_EXPIRE * 10 / 1000;
+        BaseConfigParts baseConfigParts = baseConfigPartsService
+            .findById(MEMBER_TOKEN_REFRESH_TIME);
+        if (baseConfigParts != null) {
+            duration = Long.parseLong(baseConfigParts.getCfgValue());
+        }
+        Date expireTime = new Date(now.getTime() + duration * 1000);
+
+        //之前判断是否生成过token
+        //存在一个人同时通过不同终端登录多个账号的情况
+        ThirdMemberToken entity = findById(id);
+        if (entity == null) {
+            entity = new ThirdMemberToken();
+            entity.setId(id);
+            entity.setForeignId(foreignId);
+            entity.setToken(token);
+            entity.setLoginIp(ip);
+            entity.setUpdateTime(now);
+            entity.setExpireTime(expireTime);
+            //保存token
+            save(entity);
+        } else {
+            entity.setForeignId(foreignId);
+            entity.setToken(token);
+            entity.setLoginIp(ip);
+            entity.setUpdateTime(now);
+            entity.setExpireTime(expireTime);
+            //更新token
+            update(entity);
+        }
+        TokenResponse tokenResponse = new TokenResponse();
+        tokenResponse.setToken(token);
+        tokenResponse.setExpireTime(expireTime);
+        return tokenResponse;
+    }
+
+    /**
+     * 通过有效TOKEN刷新Token，按照后台配置的时间进行刷新
+     *
+     * @param oldToken 有效的现有TOKEN
+     * @param ip       IP地址
+     * @return {@link TokenResponse}
+     */
+    public TokenResponse refreshToken(String oldToken, String ip) {
+        //生成一个token
+        String token = TokenGenerator.generateValue();
+        //当前时间
+        Date now = new Date();
+        //过期时间，默认设置为三十天，后台可调整时间，存储到BASE_CONFIG_PARRTS内，可通过数据库调整
+        Long duration = KernelConstant.THIRTY_DAY_EXPIRE / 1000;
+
+        BaseConfigParts baseConfigParts = baseConfigPartsService
+            .findById(MEMBER_TOKEN_REFRESH_TIME);
+        if (baseConfigParts != null) {
+            duration = Long.parseLong(baseConfigParts.getCfgValue());
+        }
+        Date expireTime = new Date(now.getTime() + duration * 1000);
+
+        ThirdMemberToken entity = getByToken(oldToken);
+        if (entity != null) {
+            entity.setToken(token);
+            entity.setLoginIp(ip);
+            entity.setUpdateTime(now);
+            entity.setExpireTime(expireTime);
+            //更新token
+            update(entity);
+        }
+        TokenResponse tokenResponse = new TokenResponse();
+        tokenResponse.setToken(token);
+        tokenResponse.setExpireTime(expireTime);
+        return tokenResponse;
+    }
+
+    /**
+     * 通过会员ID刷新TOKEN，按照后台配置的时间进行刷新
+     *
+     * @param id 会员ID
+     * @param ip IP地址
+     * @return {@link TokenResponse}
+     */
+    public TokenResponse refreshToken(Long id, String ip) {
+        //生成一个token
+        String token = TokenGenerator.generateValue();
+        //当前时间
+        Date now = new Date();
+        //过期时间，默认设置为三十天，后台可调整时间，存储到BASE_CONFIG_PARRTS内，可通过数据库调整
+        Long duration = KernelConstant.THIRTY_DAY_EXPIRE / 1000;
+        BaseConfigParts baseConfigParts = baseConfigPartsService
+            .findById(MEMBER_TOKEN_REFRESH_TIME);
+        if (baseConfigParts != null) {
+            duration = Long.parseLong(baseConfigParts.getCfgValue());
+        }
+        Date expireTime = new Date(now.getTime() + duration * 1000);
+
+        ThirdMemberToken entity = findById(id);
+        if (entity != null) {
+            entity.setToken(token);
+            entity.setLoginIp(ip);
+            entity.setUpdateTime(now);
+            entity.setExpireTime(expireTime);
+            //更新token
+            update(entity);
+        }
+        TokenResponse tokenResponse = new TokenResponse();
+        tokenResponse.setToken(token);
+        tokenResponse.setExpireTime(expireTime);
+        return tokenResponse;
+    }
+
+    /**
+     * 登出并且设置token无效
+     *
+     * @param id 会员ID
+     * @return true 成功; false 失败
+     */
+    public boolean logout(Long id) {
+        //生成一个token
+        String token = TokenGenerator.generateValue();
+
+        //生成一个新的token，已区分是登出状态
+        ThirdMemberToken entity = findById(id);
+        entity.setToken(token);
+        update(entity);
+
+        //保存最后一次登出时间
+        if (null != id) {
+            ThirdMember thirdMember = thirdMemberService.findById(id);
+            if (null != thirdMember && StringUtils.isNotBlank(thirdMember.getUserId())) {
+                BaseUser baseUser = baseUserService.findById(thirdMember.getUserId());
+                baseUser.setLastLogoutTime(new Date());
+                baseUserService.update(baseUser);
+            }
+        }
+        return Boolean.TRUE;
     }
 }